lab/aws/eks/services/callback-handler/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
jinja2>=3.1.4
httpx>=0.27.0
pyjwt>=2.9.0
pydantic>=2.9.0
